Move Clone Logic Into Own File
Move package cloning into a separate file before adding support for
clone from registered repositories.
